# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Pulse API
  version: 1.0.0
  description: |
    Integration health monitoring for Lakeview Health Partners (a fictional healthcare
    organisation). All data is synthetic; upstream vendor systems are simulated.

    This document is hand-written and kept in step with `docs/plan/04-api-spec.md`. CI parses it
    and asserts that every route below exists in the codebase, so drift fails the build rather
    than quietly accumulating.

    **Authentication.** Every `/api/v1/*` route except `/health` requires a session cookie issued
    by Auth.js (`POST /api/auth/callback/credentials`). Roles are ordered
    `VIEWER < OPS < ADMIN`; a route documented as `OPS` accepts OPS and ADMIN.

    **Errors.** Every failure returns the same envelope with one of
    `UNAUTHORIZED | FORBIDDEN | NOT_FOUND | VALIDATION | CONFLICT | INTERNAL`.

    **Pagination.** List routes take `cursor` and `limit` and return
    `{ data, nextCursor }`. Pass `withTotal=1` to additionally get `total`; it is opt-in because
    it costs an extra COUNT query.
servers:
  - url: /api/v1
    description: Application API
...
security:
  - sessionCookie: []
paths:
...
  /connectors:
    get:
      tags: [Connectors]
      summary: List connectors with 24h sparkline data
      responses:
        "200":
          description: Connector list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Connector" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /connectors/{key}:
    parameters:
      - $ref: "#/components/parameters/ConnectorKey"
    get:
      tags: [Connectors]
      summary: Connector detail
      description: Connector, ten most recent sync runs, 24h of health snapshots, open incident.
      responses:
        "200":
          description: Connector detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  connector: { $ref: "#/components/schemas/Connector" }
                  recentRuns:
                    type: array
                    items: { $ref: "#/components/schemas/SyncRun" }
                  snapshots:
                    type: array
                    items: { $ref: "#/components/schemas/HealthSnapshot" }
                  openIncident:
                    oneOf:
                      - $ref: "#/components/schemas/Incident"
                      - type: "null"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Connectors]
      summary: Update connector settings
      description: "Role: ADMIN. Writes an AuditEntry (`connector.update`)."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                paused: { type: boolean }
                syncIntervalSec: { type: integer, minimum: 10, nullable: true }
      responses:
        "200":
          description: Updated connector
          content:
            application/json:
              schema:
                type: object
                properties:
                  connector: { $ref: "#/components/schemas/Connector" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ValidationError" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: authjs.session-token
      description: Auth.js JWT session cookie, set by the credentials sign-in flow.
  parameters:
...
    ConnectorKey:
      name: key
      in: path
      required: true
      description: One of `ehr-fhir`, `lab-results`, `claims`, `eligibility`.
      schema: { type: string }
...
  responses:
    Unauthorized:
      description: No valid session
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Forbidden:
      description: Session lacks the required role
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    NotFound:
      description: Resource does not exist
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    ValidationError:
      description: Request body or query failed schema validation
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [UNAUTHORIZED, FORBIDDEN, NOT_FOUND, VALIDATION, CONFLICT, RATE_LIMITED, INTERNAL]
            message: { type: string }
            traceId: { type: string, nullable: true }
...
    ConnectorStatus:
      type: string
      enum: [HEALTHY, DEGRADED, DOWN, PAUSED]
    ChaosMode:
      type: string
      enum: [HEALTHY, DEGRADED, OUTAGE, TIMEOUT, RATE_LIMIT, BAD_PAYLOAD, AUTH_FAILURE]
...
    RunStatus:
      type: string
      enum: [RUNNING, SUCCEEDED, FAILED, PARTIAL]
...
    IncidentStatus:
      type: string
      enum: [OPEN, ACKNOWLEDGED, MONITORING, RESOLVED]
    IncidentSeverity:
      type: string
      enum: [CRITICAL, WARNING]
...
    Connector:
      type: object
      properties:
        id: { type: string }
        key: { type: string }
        displayName: { type: string }
        description: { type: string }
        kind:
          type: string
          enum: [poll_sync, inbound_webhook, outbound_async, request_response]
        status: { $ref: "#/components/schemas/ConnectorStatus" }
        paused: { type: boolean }
        chaosMode: { $ref: "#/components/schemas/ChaosMode" }
        syncIntervalSec: { type: integer, nullable: true }
    SyncRun:
      type: object
      properties:
        id: { type: string }
        status: { $ref: "#/components/schemas/RunStatus" }
        trigger: { type: string, enum: [schedule, manual] }
        startedAt: { type: string, format: date-time }
        finishedAt: { type: string, format: date-time, nullable: true }
        recordsFetched: { type: integer }
        recordsFailed: { type: integer }
        error: { type: string, nullable: true }
    HealthSnapshot:
      type: object
      properties:
        status: { $ref: "#/components/schemas/ConnectorStatus" }
        errorRate: { type: number, format: float }
        p95LatencyMs: { type: integer, nullable: true }
        totalCalls: { type: integer }
        failedCalls: { type: integer }
        windowStart: { type: string, format: date-time }
        windowEnd: { type: string, format: date-time }
...
    IncidentSummary:
      type: object
      description: Structured output; the schema is enforced on the model response, not parsed out of prose.
      required: [summary, probableCause, impact, suggestedSteps, confidence]
      properties:
        summary: { type: string }
        probableCause: { type: string }
        impact: { type: string }
        suggestedSteps:
          type: array
          maxItems: 5
          items: { type: string }
        confidence: { type: string, enum: [low, medium, high] }
...
    Incident:
      type: object
      properties:
        id: { type: string }
        status: { $ref: "#/components/schemas/IncidentStatus" }
        severity: { $ref: "#/components/schemas/IncidentSeverity" }
        title: { type: string }
        openedAt: { type: string, format: date-time }
        acknowledgedAt: { type: string, format: date-time, nullable: true }
        resolvedAt: { type: string, format: date-time, nullable: true }
        detectionSource: { type: string }
        aiSummaryStatus:
          type: string
          enum: [none, queued, generating, ready, failed, edited]
        aiSummary:
          description: Null until generated. An edited summary keeps the machine original under `original`.
          oneOf:
            - $ref: "#/components/schemas/IncidentSummary"
            - type: "null"
...